+++ CHANGELOG.md
@@ -1,11 +1,11 @@
-﻿# Changelog
+# Changelog
 
 All notable changes to this project will be documented in this file.
 
 The format is mainly based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
-## Unreleased
+## 5.2.0 - 2024-09-09
 
 * Added: Bring back the system measurement functions for processor, memory, storage and general system information
 * Fixed: Hide disks of remote cluster nodes when showing the Launcher PSRemoting welcome prompt